blip-29: add `swap_type` field to RFQ request message schema
This commit adds a `swap_type` field to the RFQ (Request for
Quote) request message schema. The purpose of this field is to specify
the desired swap type in the request. It enables the sender of the
request message to clearly define whether they are seeking to exchange a
Tap asset for BTC or vice versa. The addition of the `swap_type` field
enhances the versatility of the RFQ request message.