blip-29: add fields `in_asset_id` and `out_asset_id` to `tap_rfq`
Refactors `tap_rfq` to support explicit asset identification for both
incoming and outgoing assets in RFQ requests. Previously, `tap_rfq`
implied BTC as one of the assets in a swap, limiting asset flexibility.
The introduction of `in_asset_id` and `out_asset_id` enables clear
directionality for swaps, indicating the asset to send and to receive.

This update mandates that either `in_asset_id`, or `out_asset_id`, or
both must be specified. If one is omitted, BTC is assumed as the default
asset for the unspecified direction.